pyproject.toml
[project]
name = "mikrotik-backend"
version = "0.1.0"
description = "MikroTik Hotspot ISP Management Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "email-validator>=2.1.0",
    "pillow>=10.3.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "aiofiles>=23.2.1",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
